refactor: move BaseModule to module types
Reorganized imports to source BaseModule from module types instead of processor.
