+++ package.json
@@ -58,6 +58,7 @@
     "@cspell/dict-software-terms": "4.1.10",
     "@cspell/dict-typescript": "3.1.9",
     "@eslint/js": "^9.12.0",
+    "@hono/node-server": "^1.13.2",
     "@mswjs/data": "0.16.2",
     "@types/jest": "29.5.12",
     "@types/jsdom": "21.1.6",
